425 bug available soon text aligned weird sometimes (#476)
* added whiteSpace adjustments to RoomBoxHeading style line 89

* implemented test but doesnt work

* Updated test, still does not work

* Made a new file for RoomAvailabilityBox component to test the room directly, implemented test that passes

* removed RoomAvailabilityBox component declared in BuildingDrawer

* removed package-lock.json from the root folder
