+++ frontend/views/BuildingDrawer.tsx
@@ -1,5 +1,3 @@
-import { RoomStatus } from "@common/types";
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
 import CloseIcon from "@mui/icons-material/Close";
 import { Typography, TypographyProps, useMediaQuery } from "@mui/material";
 import Box, { BoxProps } from "@mui/material/Box";
@@ -13,12 +11,10 @@ import { DesktopDatePicker } from "@mui/x-date-pickers/DesktopDatePicker";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { TimePicker } from "@mui/x-date-pickers/TimePicker";
 import Image, { ImageProps } from "next/image";
-import Link from "next/link";
 import React from "react";
 
 import Button from "../components/Button";
 import useBuildingStatus from "../hooks/useBuildingStatus";
-import useRoom from "../hooks/useRoom";
 import {
   selectCurrentBuilding,
   setCurrentBuilding,
@@ -26,6 +22,7 @@ import {
 import { selectDatetime, setDatetime } from "../redux/datetimeSlice";
 import { useDispatch, useSelector } from "../redux/hooks";
 import toSydneyTime from "../utils/toSydneyTime";
+import RoomAvailabilityBox from "./RoomAvailabilityBox";
 
 const AppBox = styled(Box)(({ theme }) => ({
   boxShadow: "none",
@@ -63,37 +60,6 @@ const RoomBox = styled(Box)<BoxProps>(({ theme }) => ({
   padding: theme.spacing(0.5),
 }));
 
-const IndiviRoomBox = styled(Box)<BoxProps>(({ theme }) => ({
-  display: "flex",
-  flexDirection: "row",
-  justifyContent: "space-between",
-  alignItems: "center",
-  borderRadius: 10,
-  height: 70,
-  fontSize: 20,
-  fontWeight: 500,
-  backgroundColor: "#FFFFFF",
-  color: "black",
-  padding: theme.spacing(2, 2, 2, 3),
-  margin: theme.spacing(1.5, 1),
-  "&:hover": {
-    border: "1px solid",
-    borderColor: theme.palette.primary.main,
-    cursor: "pointer",
-  },
-}));
-
-const RoomBoxHeading = styled(Typography)<TypographyProps>(({ theme }) => ({
-  fontSize: 16,
-  fontWeight: 500,
-}));
-
-const RoomBoxSubheading = styled(Typography)<TypographyProps>(({ theme }) => ({
-  fontSize: 12,
-  fontWeight: 400,
-  paddingTop: 1,
-}));
-
 export const drawerWidth = 400;
 const drawerWidthMobile = "100%";
 
@@ -105,7 +71,9 @@ const BuildingDrawer: React.FC<{ open: boolean }> = ({ open }) => {
   const theme = useTheme();
   const isMobile = useMediaQuery(theme.breakpoints.down("sm"));
 
-  if (!building || !open) return <></>;
+  if (!building || !open) {
+    return <></>;
+  }
 
   const onClose = () => dispatch(setCurrentBuilding(null));
 
@@ -120,85 +88,6 @@ const BuildingDrawer: React.FC<{ open: boolean }> = ({ open }) => {
       }}
     />
   );
-
-  const RoomAvailabilityBox: React.FC<{
-    roomNumber: string;
-    roomStatus: RoomStatus;
-  }> = ({ roomNumber, roomStatus }) => {
-    const date = new Date(roomStatus.endtime);
-    const hoursMinutes = date.toLocaleTimeString("en-AU", {
-      hour: "2-digit",
-      minute: "2-digit",
-      hour12: true,
-    });
-
-    const roomStatusColor = {
-      free: "#2AA300",
-      busy: "#D30000",
-      soon: "#ffa600",
-    };
-
-    const roomStatusMessage = {
-      free: "Available",
-      busy: "Unavailable",
-      soon: "Available Soon",
-    };
-
-    const untilMessage = {
-      free: hoursMinutes == "Invalid Date" ? "" : "until " + hoursMinutes,
-      busy: hoursMinutes == "Invalid Date" ? "" : "until " + hoursMinutes,
-      soon: "at " + hoursMinutes,
-    };
-
-    const { room } = useRoom(`${building.id}-${roomNumber}`);
-
-    return (
-      <Link href={`/room/${building.id}-${roomNumber}`}>
-        <IndiviRoomBox>
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              alignItems: "flex-start",
-              paddingRight: 1,
-            }}
-          >
-            <RoomBoxHeading>{roomNumber}</RoomBoxHeading>
-            <RoomBoxSubheading>{!room ? "" : room.name}</RoomBoxSubheading>
-          </Box>
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "row",
-              justifyContent: "space-around",
-              alignItems: "center",
-            }}
-          >
-            <Box
-              sx={{
-                display: "flex",
-                flexDirection: "column",
-                alignItems: "flex-end",
-                paddingRight: 1,
-              }}
-            >
-              <RoomBoxHeading
-                sx={{ color: roomStatusColor[roomStatus.status] }}
-              >
-                {roomStatusMessage[roomStatus.status]}
-              </RoomBoxHeading>
-              <RoomBoxSubheading
-                sx={{ color: roomStatusColor[roomStatus.status] }}
-              >
-                {untilMessage[roomStatus.status]}
-              </RoomBoxSubheading>
-            </Box>
-            <ChevronRightIcon style={{ color: "grey" }} />
-          </Box>
-        </IndiviRoomBox>
-      </Link>
-    );
-  };
 
   return (
     <Drawer
@@ -290,6 +179,7 @@ const BuildingDrawer: React.FC<{ open: boolean }> = ({ open }) => {
                 key={roomNumber}
                 roomNumber={roomNumber}
                 roomStatus={rooms[roomNumber]}
+                buildingId={building.id}
               />
             ))
           ) : (
